Move out user facing course registration aspects to separate cog
